+++ python/GafferUI/GadgetWidget.py
@@ -55,9 +55,9 @@ class GadgetWidget( GafferUI.GLWidget ) :
 	## The gadget may either be a ViewportGadget in which case it will be used in a call
 	# to setViewportGadget, otherwise a suitable viewport will be created and the gadget will
 	# be placed within it.
-	def __init__( self, gadget=None, bufferOptions=set(), **kw ) :
+	def __init__( self, gadget=None, **kw ) :
 
-		GafferUI.GLWidget.__init__( self, bufferOptions, **kw )
+		GafferUI.GLWidget.__init__( self, **kw )
 
 		self._qtWidget().setFocusPolicy( QtCore.Qt.ClickFocus )