+++ python/GafferUI/GraphEditor.py
@@ -53,12 +53,7 @@ class GraphEditor( GafferUI.Editor ) :
 		viewportGadget.setPreciseMotionAllowed( False )
 		viewportGadget.setMaxPlanarZoom( imath.V2f( 25 ) )
 
-		self.__gadgetWidget = GafferUI.GadgetWidget(
-			gadget = viewportGadget,
-			bufferOptions = set( [
-				GafferUI.GLWidget.BufferOptions.Double,
-			] ),
-		)
+		self.__gadgetWidget = GafferUI.GadgetWidget( gadget = viewportGadget )
 
 		GafferUI.Editor.__init__( self, self.__gadgetWidget, scriptNode, **kw )