+++ Changes.md
@@ -1,10 +1,19 @@
 1.x.x.x (relative to 1.2.x.x)
 =======
 
+API
+---
+
+- ViewportGadget : A post-process shader can now be applied to any layer, not just the main one.
+
 Breaking Changes
 ----------------
 
-- GLWidget : A GL context is no longer available in `_resize()`.
+- GLWidget :
+  - A GL context is no longer available in `_resize()`.
+  - Removed `BufferOptions.Double`.
+- GadgetWidget : Removed `bufferOptions` constructor argument.
+- ViewportGadget : Changed function signature for `setPostProcessShader()` and `getPostProcessShader()`.
 
 1.2.x.x (relative to 1.2.1.1)
 =======